pyproject.toml
[tool.poetry]
name = "consumer"
version = "0.1.0"
description = "mqtt consumer app"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
paho-mqtt = "^1.5.1"
upsidedown = "^0.4"

[tool.poetry.dev-dependencies]
pytest = "^6.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
